Added player direction checks
Created a variable system that would determine the players direction by the if the players velocity was positive or negative. Using this once shooting is added the player will not be limited to only shooting right but instead can shoot in the direction they're looking in. I applied this to the player dash so now the player wont only be able to dash right