Added Melee and Redid Audio
Added melee by raycasting a circle infront of the player. This deals damage to enemies in the circle (Currently cant tell if melee is occuring so ive used the coin sound)
Redid how audio is done by using 1 audio source and changing the clip it uses when needed.
Edited the controls page to show shooting and melee
Idk probably cried alot